Allow user to remove orphaned keys
This closes #19, by making it possible to remove the orphaned
keys that might remain after generating a new one.

If there are two keys present before things start then one
will be removed as part of the generation of the new one - in that
case we'll see an errors when we try to remove that a second time,
but that is harmless.